fix: restore green build on modern toolchains
- GraphicsWindowX11: explicit casts for Xinerama screen_number
  (-Werror=sign-conversion fires when libxinerama-dev is present)
- ffmpeg plugin: migrate stale <osg/Notify>/<osg/Timer> includes missed
  by the header reorganization (plugin still targets pre-5.0 FFmpeg API;
  disable with -DBUILD_OSG_PLUGIN_FFMPEG=0 on modern systems)
- osghangglide: [[maybe_unused]] on write-only dbradius
  (-Werror=unused-but-set-global, clang 23)
- commit the thirdparty/CMakeLists.txt stub (original was gitignored and
  lost; nothing references thirdparty targets, but configure needs the
  add_subdirectory to resolve) and un-ignore it

+++ src/osgViewer/platform/GraphicsWindowX11.cpp
@@ -510,25 +510,25 @@ GraphicsWindowX11::checkAndSendEventFullScreenIfNeeded( Display* display,
             {
                 if( xi[i].x_org < minx )
                 {
-                    span[LeftMost] = xi[i].screen_number;
+                    span[LeftMost] = static_cast<uint32_t>( xi[i].screen_number );
                     minx           = xi[i].x_org;
                 }
 
                 if( xi[i].x_org > maxx )
                 {
-                    span[RightMost] = xi[i].screen_number;
+                    span[RightMost] = static_cast<uint32_t>( xi[i].screen_number );
                     maxx            = xi[i].x_org;
                 }
 
                 if( xi[i].y_org < miny )
                 {
-                    span[TopMost] = xi[i].screen_number;
+                    span[TopMost] = static_cast<uint32_t>( xi[i].screen_number );
                     miny          = xi[i].y_org;
                 }
 
                 if( xi[i].y_org > maxy )
                 {
-                    span[BottomMost] = xi[i].screen_number;
+                    span[BottomMost] = static_cast<uint32_t>( xi[i].screen_number );
                     maxy             = xi[i].y_org;
                 }
             }
